Upgraded Guava to unmodified jdk5-backport-v17.0-compatibility
This simply copies the Guava source for jdk5-backport-v17.0-compatibility
straight from the github repository into this one.

See https://github.com/google/guava.git

Additional commits will be made which will allow this to compile
on Android.

+++ guava-tests/test/com/google/common/primitives/UnsignedLongsTest.java
@@ -21,13 +21,13 @@ import com.google.common.annotations.GwtIncompatible;
 import com.google.common.collect.testing.Helpers;
 import com.google.common.testing.NullPointerTester;
 
+import junit.framework.TestCase;
+
 import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Random;
-
-import junit.framework.TestCase;
 
 /**
  * Tests for UnsignedLongs
@@ -98,10 +98,10 @@ public class UnsignedLongsTest extends TestCase {
         new long[] {},
         new long[] {LEAST},
         new long[] {LEAST, LEAST},
-        new long[] {LEAST, 1},
-        new long[] {1},
-        new long[] {1, LEAST},
-        new long[] {GREATEST, GREATEST - 1},
+        new long[] {LEAST, (long) 1},
+        new long[] {(long) 1},
+        new long[] {(long) 1, LEAST},
+        new long[] {GREATEST, GREATEST - (long) 1},
         new long[] {GREATEST, GREATEST},
         new long[] {GREATEST, GREATEST, GREATEST});
